// apps/stroltm/ui/src/api/generated/api.ts
export interface Config {
  /**
   *
   * @type {boolean}
   * @memberof Config
   */
  disableWatchChanges?: boolean;
  /**
   *
   * @type {{ [key: string]: { [key: string]: ConfigServiceTask; }; }}
   * @memberof Config
   */
  services?: { [key: string]: { [key: string]: ConfigServiceTask } };
  /**
   *
   * @type {Array<string>}
   * @memberof Config
   */
  tags?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof Config
   */
  timezone?: string;
}
/**
 *
 * @export
 * @interface ConfigServiceTask
 */
export interface ConfigServiceTask {
  /**
   *
   * @type {{ [key: string]: ConfigServiceTaskDestination; }}
   * @memberof ConfigServiceTask
   */
  destinations?: { [key: string]: ConfigServiceTaskDestination };
  /**
   *
   * @type {Array<ConfigServiceTaskNotification>}
   * @memberof ConfigServiceTask
   */
  notifications?: Array<ConfigServiceTaskNotification>;
  /**
   *
   * @type {ConfigServiceTaskSchedule}
   * @memberof ConfigServiceTask
   */
  schedule?: ConfigServiceTaskSchedule;
  /**
   *
   * @type {ConfigServiceTaskSource}
   * @memberof ConfigServiceTask
   */
  source?: ConfigServiceTaskSource;
  /**
   *
   * @type {Array<string>}
   * @memberof ConfigServiceTask
   */
  tags?: Array<string>;
}
// ...
export interface ManagerPreparedInstance {
  /**
   *
   * @type {Config}
   * @memberof ManagerPreparedInstance
   */
  config?: Config;
  /**
   *
   * @type {boolean}
   * @memberof ManagerPreparedInstance
   */
  isOnline?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerPreparedInstance
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerPreparedInstance
   */
  proxyName?: string;
  /**
   *
   * @type {ManagerStatus}
   * @memberof ManagerPreparedInstance
   */
  taskStatus?: ManagerStatus;
}
/**
 *
 * @export
 * @interface ManagerStatus
 */
export interface ManagerStatus {
  /**
   *
   * @type {string}
   * @memberof ManagerStatus
   */
  lastChangedAt?: string;
  /**
   *
   * @type {Array<ManagerTaskItem>}
   * @memberof ManagerStatus
   */
  tasks?: Array<ManagerTaskItem>;
}
/**
 *
 * @export
 * @interface ManagerTaskItem
 */
export interface ManagerTaskItem {
  /**
   *
   * @type {boolean}
   * @memberof ManagerTaskItem
   */
  isRunning?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  lastEndedAt?: string;
  /**
   *
   * @type {OperationType}
   * @memberof ManagerTaskItem
   */
  operation?: OperationType;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  serviceName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  startedAt?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  taskName?: string;
  /**
   *
   * @type {TriggerType}
   * @memberof ManagerTaskItem
   */
  trigger?: TriggerType;
}